+++ scripts/call-recording_download.rb
@@ -10,8 +10,8 @@ require 'pp'
 config = RingCentralSdk::REST::Config.new.load_dotenv
 
 client = RingCentralSdk::REST::Client.new
-client.app_config(config.app)
-client.authorize_user(config.user)
+client.app_config config.app
+client.authorize_user config.user
 
 def get_recordings(client)
   # Retrieve voice call log records with recordings